Add files via upload
Created a Person class that has a constructor that takes name and surname parameters and contains a showFullName() method that displays the person's first and last names.
From the Person class, the Student class is inherited, the constructor of which, in addition to name and surname, takes the parameter year (the year of entering the university).
In the Student class, overrided the showFullName(midleName) method to display not only the first and last name, but also the middle name (midleName) of the student.
Also, in the Student class, implemented the showCourse() method, which will display the student's current course (from 1 to 6). The value of the course will be determined as the difference between the current year (to determine independently) and the year of admission to the university year.